Changed structure of company to store lists of values instead of lists of HistoricalDataResponse class Hopefully this will be easier to work with and faster

+++ Trading/Database.cs
@@ -12,16 +12,20 @@ namespace Trading
     public class Database
     {
         public LiteDatabase DB;
-        public Database(string filename)
+        public Database(string filename, bool newdb)
         {
-            if (File.Exists(filename))
-                File.Delete(filename);
+            if (newdb)
+            {
+                if (File.Exists(filename))
+                    File.Delete(filename);
+            }
             DB = new LiteDatabase(@filename);
 
         }
-        public Company DownloadSymbol(string symbol)
+        public void AddSymbol(Company c)
         {
-            return IEXData.DownloadSymbol(symbol, IEXData.HistoryType.FiveYear);
+            var col = DB.GetCollection<Trading.Company>("data");
+            col.Insert(c);
         }
         public Company GetSymbolFromDB(string symbol)
         {
@@ -29,6 +33,26 @@ namespace Trading
 
             var c = col.FindOne(x => x.Symbol.Equals(symbol, StringComparison.CurrentCultureIgnoreCase));
             return (Company)c;
+        }
+
+        public void GetAllData(IEXData.HistoryType ht, int num)
+        {
+           List<Trading.SymbolData> symbols = Trading.IEXData.DownloadSymbolList();
+
+           int cnt = symbols.Count;
+            if (num > 0)
+                cnt = num;
+
+
+            for (int i = 0; i < cnt; i++)
+            {
+                Trading.SymbolData symData = symbols[i];
+                Console.WriteLine((i+1).ToString() + "/" + symbols.Count.ToString() + "\t" + symData.symbol  );
+                Trading.Company c = Trading.IEXData.DownloadSymbol(symData.symbol, ht);
+                if (c.Id != -1)
+                    this.AddSymbol(c);
+                //System.Threading.Thread.Sleep(20);
+            }
         }
     }
 }
